% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{add_edge}
\alias{add_edge}
\title{add_edge}
\usage{
add_edge(graph, from, to, verbose = FALSE, ...)
}
\arguments{
\item{graph}{A dot diagram created from `DiagrammeR::grViz`}

\item{from}{The node or nodes to draw the edge from. Multiple nodes can be specified using `c()`}

\item{to}{The node or nodes to draw the edge to. Multiple nodes can be specified using `c()`}

\item{verbose}{Prints out the dot diagram in html format, Default: FALSE}

\item{...}{Parameters passed to \code{\link[DiagrammeR]{grViz}}}
}
\value{
A graph object of class grViz and htmlwidget
}
\description{
Adds an edge or edges to connect two or more nodes together
}
\details{
Adds an edge or edges to connect two or more nodes together
}
\seealso{
\code{\link[DiagrammeR]{grViz}}
}
