% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{add_node}
\alias{add_node}
\title{add_node}
\usage{
add_node(graph, name, stringWidth = 50, verbose = FALSE, ...)
}
\arguments{
\item{graph}{A dot diagram created from `DiagrammeR::grViz`}

\item{name}{The names given to the node or nodes. Multiple nodes can be specified using `c()`}

\item{stringWidth}{Positive integer giving the line width (in number of characters), Default: 50}

\item{verbose}{Prints out the dot diagram in html format, Default: FALSE}

\item{...}{Parameters passed to \code{\link[DiagrammeR]{grViz}}}
}
\value{
A graph object of class grViz and htmlwidget
}
\description{
Adds a node or nodes to a dot diagram
}
\details{
Adds a node or nodes to a dot diagram
}
\seealso{
\code{\link[DiagrammeR]{grViz}}
}
