% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_labels}
\alias{rotate_labels}
\title{rotate_labels}
\usage{
rotate_labels(graph, nodes, angle = -90, moveX = 0, moveY = 0)
}
\arguments{
\item{graph}{A dot diagram created from `DiagrammeR::grViz`}

\item{nodes}{An integer or numeric vector containing the node number(s) indicating which labels to rotate}

\item{angle}{The angle for which the labels should be rotated, Default: -90}

\item{moveX}{A numeric value to shift the label along the x-axis}

\item{moveY}{A numeric value to shift the label along the y-axis}
}
\value{
A graph object of class grViz and htmlwidget with rotated labels
}
\description{
Rotates the labels of the specified nodes from the grViz and htmlwidget object
}
\details{
Rotates the labels of the specified nodes from the grViz and htmlwidget object
}
\seealso{
\code{\link[htmltools]{HTML}}
 \code{\link[htmlwidgets]{prependContent}}, \code{\link[htmlwidgets]{onStaticRenderComplete}}
}
