% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_consort_as_pdf.R
\name{save_consort_as_pdf}
\alias{save_consort_as_pdf}
\title{save_consort_as_pdf}
\usage{
save_consort_as_pdf(graph, outdir, moveX = 0, moveY = 0)
}
\arguments{
\item{graph}{A dot diagram created from `DiagrammeR::grViz`}

\item{outdir}{Output directory to save the pdf file}

\item{moveX}{A numeric value to shift the label along the x-axis}

\item{moveY}{A numeric value to shift the label along the y-axis}
}
\value{
Saves the consort diagram as a pdf document
}
\description{
Saves the consort diagram built in tidyconsort to a pdf file
}
\details{
Saves the consort diagram built in tidyconsort to a pdf file
}
\seealso{
\code{\link[htmlwidgets]{saveWidget}}
 \code{\link[xml2]{read_xml}}, \code{\link[xml2]{xml_text}}, \code{\link[xml2]{xml_find_all}}, \code{\link[xml2]{xml_ns}}, \code{\link[xml2]{xml_attr}}, \code{\link[xml2]{write_xml}}
 \code{\link[DiagrammeRsvg]{export_svg}}
 \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_remove}}, \code{\link[stringr]{str_match}}
 \code{\link[rsvg]{rsvg}}
}
