% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_edge_parameters}
\alias{set_edge_parameters}
\title{set_edge_parameters}
\usage{
set_edge_parameters(graph, verbose = FALSE, ...)
}
\arguments{
\item{graph}{A dot diagram created from `DiagrammeR::grViz`}

\item{verbose}{Prints out the dot diagram in html format, Default: FALSE}

\item{...}{Parameters passed to \code{\link[DiagrammeR]{grViz}}}
}
\value{
A graph object of class grViz and htmlwidget
}
\description{
Sets multiple parameters of an edge until those parameters are altered.
}
\details{
Sets multiple parameters of an edge until those parameters are altered.
}
\seealso{
\code{\link[DiagrammeR]{grViz}}
}
