% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyconsort-package.R
\docType{package}
\name{tidyconsort-package}
\alias{tidyconsort}
\alias{tidyconsort-package}
\title{tidyconsort: An Interface to the DiagrammeR Package to Create Consort Diagrams}
\description{
The goal of tidyconsort is to make creating consort diagrams easier with the DiagrammeR package.
}
\author{
\strong{Maintainer}: Brian C. Helsel \email{bhelsel@kumc.edu}

}
\keyword{internal}
