% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{use_template_1}
\alias{use_template_1}
\title{use_template_1}
\usage{
use_template_1(
  eligible = 0,
  randomized = 0,
  timepoints = 2,
  labels = c("Timepoint 1", "Timepoint 2"),
  descriptions = c("Baseline Testing", "Post-intervention"),
  intervention = "Intervention",
  intervention_n = 0,
  control = "Control",
  control_n = 0,
  verbose = FALSE
)
}
\arguments{
\item{eligible}{Number of study participants eligible for the study, Default: 0}

\item{randomized}{Number of study participants randomized during the study, Default: 0}

\item{timepoints}{The number of time points to include in the CONSORT diagram (range: 2-4), Default: 2}

\item{labels}{The labels for the time points (e.g., baseline, 6 months, etc.), Default: c("Timepoint 1", "Timepoint 2")}

\item{descriptions}{A description of the trial phase completed (e.g., active intervention, no contact follow-up), Default: c("Baseline Testing", "Post-intervention")}

\item{intervention}{A name or descriptor for the intervention group, Default: 'Intervention'}

\item{intervention_n}{The number of individuals completing each trial phase (this should match the length of the descriptions), Default: 0}

\item{control}{A name or descriptor for the intervention group (e.g., placebo), Default: 'Control'}

\item{control_n}{The number of individuals completing each trial phase (this should match the length of the descriptions), Default: 0}

\item{verbose}{Prints out the dot diagram in html format, Default: FALSE}
}
\value{
A graph object of class grViz and htmlwidget
}
\description{
A basic example of a CONSORT diagram created with tidyconsort
}
\details{
A basic example of a CONSORT diagram created with tidyconsort. The
    default is a two-arm randomized trial with changeable parameters to add up
    to 4 time points.
}
\examples{
\dontrun{
if(interactive()){
  use_template_1()
 }
}
}
